@@ -6,9 +6,9 @@ authors = [{ name = "Sebastian Waldvogel", email = "sebastian@waldvogels.de" }]
 license = { text = "MIT" }
 readme = "README.md"
 
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
-    "pye3dc==0.9.3",
+    "pye3dc==0.10.0",
     "tzlocal==5.3.1",
     "jsonargparse==4.42.0",
     "rich-argparse==1.7.2",
